requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfminer.six>=20221105
lxml>=4.9.0
